Safely write to md5 cache file
The currently used technique with appending to file could lead
to md5 sum being written twice into the file in case two devices
are being bootstrapped simultaneously. Which actually happend in my
case when I've tried that with a couple Arista switches.

Replaced that code with write-replace pattern, which is a safe way to
do the same thing. For details see that blog post:
https://blog.gocept.com/2013/07/15/reliable-file-updates-with-python/